pyproject.toml
[tool.poetry]
name = "uno"
version = "0.9.0"
description = "Dynamic site-to-site VPNs with WireGuard, frrouting, and RTI Connext DDS"
authors = ["Andrea Sorbini <as@mentalsmash.org>"]
readme = "README.md"
license = "Apache-2.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: POSIX :: Linux"
]
packages = [
  { include = "uno" }
]

[tool.poetry.dependencies]
google-api-python-client = "^2.124.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
Jinja2 = "^3.1.2"
pygments = "^2.17.2"
pymdown-extensions = "^10.7.1"
pytest = "^8.1"
python = "^3.9"
python-gnupg = "^0.5.2"
PyYAML = "^6.0"
markdown = "^3.6"
matplotlib = "^3.8.3"
networkx = "^3.2.1"
sdnotify = "^0.3.2"
tabulate = "^0.9.0"
termcolor = "^2.4.0"

[tool.poetry.group.connext]
optional = true

[tool.poetry.group.connext.dependencies]
"rti.connext" = "^7.1.0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
docutils = "^0.18.1"
m2r = "^0.3.1"
mistune = "^0.8.4"
sphinx-autodoc-typehints = "^1.25.2"
sphinx-rtd-theme = "^1.3.0"
sphinxcontrib-programoutput = "^0.17"
sphinxemoji = "^0.2.0"
sphinx-tabs = "^3.4.4"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.7"
pre-commit = "^3.7.0"

[tool.pytest.ini_options]
addopts = [
  "--import-mode=importlib",
]

[tool.poetry.scripts]
uno = 'uno.cli.uno:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
indent-width = 2
target-version = "py39"

# [tool.ruff.lint]
# ignore = ["E741"]